Fixed RuntimeError
RuntimeError: Expected object of scalar type Float but got scalar type Double for sequence element 2 in sequence argument at position #1 'tensors'

+++ kits19cnn/io/custom_transforms.py
@@ -90,7 +90,7 @@ class RandomResizedCropTransform(AbstractTransform):
 
         data_dict[self.data_key] = data
         if seg is not None:
-            data_dict[self.label_key] = seg
+            data_dict[self.label_key] = seg.astype(np.float32)
 
         return data_dict
